+++ purchase/models.py
@@ -10,7 +10,7 @@ class Purchase(models.Model):
     '''
     purchase_id = models.CharField(max_length=8, primary_key=True, blank=True, verbose_name='采购编号')
     deliver_id = models.ForeignKey(Deliver, on_delete=models.CASCADE,related_name="purchase_deliver")
-    purchase_num = models.IntegerField(verbose_name='采购数量')
+    purchase_num = models.FloatField(verbose_name='采购数量')
     purchase_time = models.DateTimeField(verbose_name='采购时间')
     purchase_price = models.FloatField(max_length=8, verbose_name='采购的价格')
 
@@ -21,7 +21,7 @@ class Supplier(models.Model):
     supplier_id = models.CharField(max_length=8, primary_key=True, blank=True, verbose_name='供应商编号')
     supplier_add = models.TextField(verbose_name='供应商地址')
     supplier_name = models.CharField(max_length=20, blank=True, null=True, verbose_name='供应商名字')
-    supplier_charge_phone = models.IntegerField(verbose_name='供应商负责人电话')
+    supplier_charge_phone = models.PositiveBigIntegerField(verbose_name='供应商负责人电话')
     supplier_charge_name = models.CharField(max_length=20, blank=True, null=True, verbose_name='供应商负责人名字')
 
 class PurchaseDetail(models.Model):
